Fixing some bugs and getting tests working again
