+++ .vscode/launch.json
@@ -13,14 +13,14 @@
             ],
             "preLaunchTask": "Build App",
             "program": "${workspaceFolder}/app/src/main.ts",
-            "args": ["sync"],
+            "args": ["token"],
             "outFiles": [
                 "${workspaceFolder}/app/bin/**/*.js",
             ],
             "env": {
                 "NODE_NO_WARNINGS": "1"
             },
-            "envFile": "${workspaceFolder}/.vscode/private.env"
+            "envFile": "${workspaceFolder}/.vscode/test.env"
         }, {
             "name": "Run API Tests",
             "type": "node",
@@ -34,6 +34,9 @@
             "console": "integratedTerminal",
             "internalConsoleOptions": "neverOpen",
             "cwd": "${workspaceFolder}/app",
+            "env": {
+                "NODE_NO_WARNINGS": "1"
+            },
             "envFile": "${workspaceFolder}/.vscode/api-test.env"
         }, {
             "name": "Run All Tests",
@@ -48,7 +51,10 @@
             "console": "integratedTerminal",
             "internalConsoleOptions": "neverOpen",
             "cwd": "${workspaceFolder}/app",
-            "envFile": "${workspaceFolder}/.vscode/api-test.env"
+            "envFile": "${workspaceFolder}/.vscode/api-test.env",
+            "env": {
+                "NODE_NO_WARNINGS": "1"
+            },
         }, {
             "name": "Debug Test",
             "type": "node",
@@ -60,7 +66,7 @@
                 "--runInBand",
                 "--config", "jest.config.json",
                 //"--detectOpenHandles",
-                "test/unit/icloud.test.ts"
+                "test/unit/sync-engine.test.ts"
             ],
             "env": {
                 "NODE_NO_WARNINGS": "1"
