+++ app/jest.config.json
@@ -13,7 +13,7 @@
       "!src/app/event/**/*",
       "!src/app/error/**/*",
       "!src/lib/package.ts",
-      "!src/lib/photos-library/model photos-entity.ts"
+      "!src/lib/photos-library/model/photos-entity.ts"
     ],
     "coverageProvider": "v8",
     "coverageReporters": [
