+++ app/test/_helpers/_config.ts
@@ -7,7 +7,7 @@ export const defaultConfig = {
     password: `testPass`,
     dataDir: `/opt/icloud-photos-library`,
     port: 80,
-    maxRetries: Infinity,
+    maxRetries: 10,
     downloadThreads: 5,
     schedule: `0 2 * * *`,
     enableCrashReporting: false,
@@ -21,7 +21,7 @@ export const defaultConfig = {
     suppressWarnings: false,
     exportMetrics: false,
     metadataRate: [Infinity, 0],
-    networkCapture: true,
+    networkCapture: false,
 } as iCPSAppOptions;
 
 export const trustToken = `HSARMTKNSRVWFlaje+uXQGSn7KqxCHcGvTYF5r7F6j8wnFOsemu+g20/1mNJmE+hfNgaB09Bt3RDLzU4kWCsjQUvEsv3C//DrFvldh26AGf9sSv8QyRkUFNJjSHH7fHHiWaZRlWIZk9viZVnxiAh+NE/cA9ZGpbshwkVErwD5/cHN+Ek69ufbIS0O5P0eA==SRVX`;
